Read and Create Tests

+++ webtoonQuiz/src/main/java/com/webtoonquiz/repo/OxQuizRepository.java
@@ -2,6 +2,6 @@ package com.webtoonquiz.repo;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.webtoonquiz.model.OxQuiz;
-public interface OxQuizRepository extends JpaRepository<OxQuiz , Integer> {
+public interface OxQuizRepository extends JpaRepository<OxQuiz , Long> {
 
 }
